refactor: 💡 separate TCP stream handling into its own module for improved organization

+++ server/src/app.rs
@@ -1,21 +1,19 @@
 mod command_processor;
+mod tcp_stream_handler;
 
 use std::{
-    fs::{self, remove_dir_all},
+    fs::{create_dir_all, remove_dir_all},
     net::TcpListener,
     path::Path,
 };
 
-use serde_json::json;
-use shared::{
-    app,
-    mmp::{self},
-};
+use shared::{app, mmp};
+use tcp_stream_handler::TcpStreamHandler;
 
 pub fn run() -> anyhow::Result<()> {
     let app_config = app::Config::new()?;
 
-    let tcp_listener = TcpListener::bind(app_config.server_addr)?;
+    let tcp_listener = TcpListener::bind(app_config.server_addr.clone())?;
 
     for tcp_stream in tcp_listener.incoming() {
         match tcp_stream {
@@ -32,33 +30,12 @@ pub fn run() -> anyhow::Result<()> {
                 ));
 
                 if !download_dir_path.exists() {
-                    fs::create_dir(&download_dir_path)?;
+                    create_dir_all(&download_dir_path)?;
                 }
 
-                let input_file_path_without_ext = download_dir_path.join("input");
-                let output_file_path_without_ext = download_dir_path.join("output");
-
-                let received_packet = mmp_stream.receive_packet(&input_file_path_without_ext)?;
-
-                println!("Received packet: {:?}", received_packet);
-
-                let input_file_path: &Path = received_packet.payload.media_file_path.as_ref();
-                let request_json: app::Request = serde_json::from_value(received_packet.json.data)?;
-
-                let output_file_path = command_processor::CommandProcessor::process(
-                    request_json.command,
-                    input_file_path,
-                    &output_file_path_without_ext,
-                )?;
-
-                let response_packet = mmp::Packet::new(
-                    mmp::Json::new(json!(mmp::Response::new(mmp::Status::Ok)))?,
-                    mmp::Payload::new(output_file_path.to_path_buf())?,
-                );
-
-                mmp_stream.send_packet(&response_packet)?;
-
-                remove_dir_all(&download_dir_path)?;
+                if let Err(_error) = TcpStreamHandler::handle(&mut mmp_stream, &download_dir_path) {
+                    remove_dir_all(&download_dir_path)?;
+                }
             }
             Err(e) => {
                 eprintln!("Error accepting connection: {}", e);
